amt: Modernize the plugin and split out common functionality
The MEI code is soon going to be used by a new plugin, so split it
out into libfwupdplugin like we did i2c.
