+++ libfwupdplugin/fwupdplugin.h
@@ -48,6 +48,7 @@
 #include <libfwupdplugin/fu-ihex-firmware.h>
 #include <libfwupdplugin/fu-io-channel.h>
 #include <libfwupdplugin/fu-kernel.h>
+#include <libfwupdplugin/fu-mei-device.h>
 #include <libfwupdplugin/fu-mem.h>
 #include <libfwupdplugin/fu-path.h>
 #include <libfwupdplugin/fu-plugin-vfuncs.h>
