+++ libfwupdplugin/meson.build
@@ -66,6 +66,7 @@ fwupdplugin_src = [
   'fu-efivar.c',
   'fu-udev-device.c',
   'fu-i2c-device.c',
+  'fu-mei-device.c',
   'fu-usb-device.c',
   'fu-cfi-device.c',
   'fu-hid-device.c',
@@ -147,6 +148,7 @@ fwupdplugin_headers = [
   'fu-efivar.h',
   'fu-udev-device.h',
   'fu-i2c-device.h',
+  'fu-mei-device.h',
   'fu-usb-device.h',
   'fu-hid-device.h',
 ]
