+++ plugins/amt/amt.quirk
@@ -1,0 +1,2 @@
+[MEI]
+Plugin = amt
