+++ plugins/amt/meson.build
@@ -1,10 +1,15 @@
 if get_option('plugin_amt').disable_auto_if(host_machine.system() != 'linux').allowed()
 cargs = ['-DG_LOG_DOMAIN="FuPluginAmt"']
 
+install_data('amt.quirk',
+  install_dir: join_paths(datadir, 'fwupd', 'quirks.d')
+)
+
 shared_module('fu_plugin_amt',
   fu_hash,
   sources : [
     'fu-plugin-amt.c',
+    'fu-amt-device.c',
   ],
   include_directories : [
     root_incdir,
